dbクラスをdbクラスとmemoDbクラスに分ける修正
